Slight modifications
Main is a test ground for running the code. Main can be used for the actually game that will be loaded into the game engine. My goal after setting the framework up for character movement  is to test loading levels in the game. I need to find a way to do this using either a text file or a binary file.